Добавлен компонент карточки фильма
